proxy-server/requirements.txt
flask
openai
python-dotenv 